requirements.txt
altair==5.5.0
attrs==25.3.0
black==25.1.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
-e git+ssh://git@github.com/JoeJDLaw/company_junction.git@cd9c1797dd89bcb2594497fdcd8fec0b603f6970#egg=company_junction
coverage==7.10.5
cryptography==45.0.6
duckdb==1.3.2
et_xmlfile==2.0.0
flake8==7.3.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
loguru==0.7.3
lxml==6.0.1
MarkupSafe==3.0.2
mccabe==0.7.0
more-itertools==10.7.0
mypy==1.17.1
mypy_extensions==1.1.0
narwhals==2.2.0
numpy==2.3.2
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pandas-stubs==2.3.2.250827
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pyarrow-stubs==20.0.0.20250825
pycodestyle==2.14.0
pycparser==2.22
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.2
pyinstrument==5.1.1
PyJWT==2.10.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.0
referencing==0.36.2
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
ripgrepy==2.2.0
rpds-py==0.27.1
ruff==0.12.10
simple-salesforce==1.12.9
six==1.17.0
smmap==5.0.2
streamlit==1.49.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
types-psutil==7.0.0.20250822
types-pytz==2025.2.0.20250809
types-PyYAML==6.0.12.20250822
types-setuptools==80.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vulture==2.14
watchdog==6.0.0
xlrd==2.0.2
zeep==4.3.1
